app/requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
boto3==1.35.36
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
